Extract CUOBJ tests to their own plugin test file.  Make endpoint override a universaly applied environment variable to allow all tests to run against a specific endpoint
