+++ test/gtest/plugins/meson.build
@@ -44,8 +44,13 @@ if cuda_dep.found()
     plugins_gtest_cuda_deps = [cuda_dep]
 endif
 
+plugins_gtest_sources = ['../main.cpp', '../common.cpp', 'obj_plugin.cpp']
+if cuobj_dep.found()
+    plugins_gtest_sources += 'obj_cuobj_plugin.cpp'
+endif
+
 plugins_test_exe = executable('plugins_gtest',
-    sources : ['../main.cpp', '../common.cpp', 'obj_plugin.cpp'],
+    sources : plugins_gtest_sources,
     include_directories: [nixl_inc_dirs, utils_inc_dirs, plugins_inc_dirs, '.'],
     cpp_args : plugins_gtest_cpp_flags,
     dependencies : [nixl_dep, gtest_dep, absl_strings_dep, absl_time_dep, plugin_deps,
