+++ test/gtest/plugins/obj_plugin.cpp
@@ -33,34 +33,31 @@ namespace gtest::plugins::obj {
  * These variables are required for authenticating and interacting with the S3 bucket
  * used during the tests.
  *
- * For Dell ObjectScale tests, the following additional environment variable is required:
- *       - NIXL_OBJ_ENDPOINT_OVERRIDE  (e.g. http://100.68.213.151:9020)
- *
  * Test suites:
  * - ObjTests: Standard S3 client tests (crtMinLimit = 0)
  * - ObjCrtTests: S3 CRT client tests (crtMinLimit = 5 MiB, buffer = 10 MiB)
  *                crtMinLimit is set to the S3 minimum part size (5 MiB) so that
  *                partSize is not clamped and MPU is exercised with multiple parts.
  * - ObjAccelTests: S3 Accelerated client tests (accelerated = true)
- *                  Note: Only compiled if HAVE_CUOBJ_CLIENT is defined
+ *                  Note: Defined in obj_cuobj_plugin.cpp, only compiled if
+ *                  HAVE_CUOBJ_CLIENT is defined
  * - ObjDellTests: Dell ObjectScale S3 over RDMA tests
  *                 (accelerated = true, type = dell, req_checksum = required, scheme = http)
- *                 Note: Only compiled if HAVE_CUOBJ_CLIENT is defined.
+ *                 Note: Defined in obj_cuobj_plugin.cpp, only compiled if
+ *                 HAVE_CUOBJ_CLIENT is defined.
  *                 Skipped at runtime if NIXL_OBJ_ENDPOINT_OVERRIDE is not set.
  *                 Includes DellVramXferTest (GPU memory) if HAVE_CUDA is also defined.
+ *
+ * Environment:
+ *       - NIXL_OBJ_ENDPOINT_OVERRIDE  (e.g. http://100.68.213.151:9020)
+ *         Optional for all test suites. When set, overrides the S3 endpoint for
+ *         Standard, CRT, Accel, and Dell tests alike.
  */
 
 nixl_b_params_t obj_params = {{"crtMinLimit", "0"}};
 nixl_b_params_t obj_crt_params = {{"crtMinLimit", "5242880"}}; // 5 MiB: S3 minimum part size
-nixl_b_params_t obj_accel_params = {{"accelerated", "true"}};
-nixl_b_params_t obj_dell_params = {{"accelerated", "true"},
-                                   {"type", "dell"},
-                                   {"req_checksum", "required"},
-                                   {"scheme", "http"}};
 const std::string local_agent_name = "Agent1";
 const std::string crt_agent_name = "Agent2-CRT";
-const std::string accel_agent_name = "Agent3-Accel";
-const std::string dell_agent_name = "Agent4-Dell";
 const nixlBackendInitParams obj_test_params = {.localAgent = local_agent_name,
                                                .type = "OBJ",
                                                .customParams = &obj_params,
@@ -76,25 +73,14 @@ const nixlBackendInitParams obj_crt_test_params = {.localAgent = crt_agent_name,
                                                    .syncMode =
                                                        nixl_thread_sync_t::NIXL_THREAD_SYNC_RW};
 
-const nixlBackendInitParams obj_accel_test_params = {.localAgent = accel_agent_name,
-                                                     .type = "OBJ",
-                                                     .customParams = &obj_accel_params,
-                                                     .enableProgTh = false,
-                                                     .pthrDelay = 0,
-                                                     .syncMode =
-                                                         nixl_thread_sync_t::NIXL_THREAD_SYNC_RW};
-
-const nixlBackendInitParams obj_dell_test_params = {.localAgent = dell_agent_name,
-                                                    .type = "OBJ",
-                                                    .customParams = &obj_dell_params,
-                                                    .enableProgTh = false,
-                                                    .pthrDelay = 0,
-                                                    .syncMode =
-                                                        nixl_thread_sync_t::NIXL_THREAD_SYNC_RW};
-
 class setupObjTestFixture : public setupBackendTestFixture {
 protected:
     setupObjTestFixture() {
+        const char *endpoint = std::getenv("NIXL_OBJ_ENDPOINT_OVERRIDE");
+        if (endpoint && endpoint[0] != '\0') {
+            obj_params["endpoint_override"] = endpoint;
+            obj_params["req_checksum"] = "required";
+        }
         localBackendEngine_ = std::make_shared<nixlObjEngine>(&GetParam());
     }
 };
@@ -144,6 +130,10 @@ INSTANTIATE_TEST_SUITE_P(ObjTests, setupObjTestFixture, testing::Values(obj_test
 class setupObjCrtTestFixture : public setupBackendTestFixture {
 protected:
     setupObjCrtTestFixture() {
+        const char *endpoint = std::getenv("NIXL_OBJ_ENDPOINT_OVERRIDE");
+        if (endpoint && endpoint[0] != '\0') {
+            obj_crt_params["endpoint_override"] = endpoint;
+        }
         localBackendEngine_ = std::make_shared<nixlObjEngine>(&GetParam());
     }
 };
@@ -206,146 +196,5 @@ TEST_P(setupObjCrtTestFixture, CrtQueryMemTest) {
 }
 
 INSTANTIATE_TEST_SUITE_P(ObjCrtTests, setupObjCrtTestFixture, testing::Values(obj_crt_test_params));
-
-#if defined HAVE_CUOBJ_CLIENT
-// Separate test suite for S3 Accelerated client with accelerated=true
-// Note: These tests require the cuobjclient library to be available at compile time
-class setupObjAccelTestFixture : public setupBackendTestFixture {
-protected:
-    setupObjAccelTestFixture() {
-        localBackendEngine_ = std::make_shared<nixlObjEngine>(&GetParam());
-    }
-};
-
-TEST_P(setupObjAccelTestFixture, AccelXferTest) {
-    transferHandler<DRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, accel_agent_name, accel_agent_name, false, 1);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-    transfer.resetLocalMem();
-    transfer.testTransfer(NIXL_READ);
-    transfer.checkLocalMem();
-}
-
-TEST_P(setupObjAccelTestFixture, AccelXferMultiBufsTest) {
-    transferHandler<DRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, accel_agent_name, accel_agent_name, false, 3);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-    transfer.resetLocalMem();
-    transfer.testTransfer(NIXL_READ);
-    transfer.checkLocalMem();
-}
-
-TEST_P(setupObjAccelTestFixture, AccelQueryMemTest) {
-    transferHandler<DRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, accel_agent_name, accel_agent_name, false, 3);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-
-    nixl_reg_dlist_t descs(OBJ_SEG);
-    descs.addDesc(nixlBlobDesc(nixlBasicDesc(), "test-obj-key-0"));
-    descs.addDesc(nixlBlobDesc(nixlBasicDesc(), "test-obj-key-1"));
-    descs.addDesc(nixlBlobDesc(nixlBasicDesc(), "test-obj-key-nonexistent"));
-    std::vector<nixl_query_resp_t> resp;
-    localBackendEngine_->queryMem(descs, resp);
-
-    EXPECT_EQ(resp.size(), 3);
-    EXPECT_EQ(resp[0].has_value(), true);
-    EXPECT_EQ(resp[1].has_value(), true);
-    EXPECT_EQ(resp[2].has_value(), false);
-}
-
-INSTANTIATE_TEST_SUITE_P(ObjAccelTests,
-                         setupObjAccelTestFixture,
-                         testing::Values(obj_accel_test_params));
-
-/**
- * @brief Test fixture for Dell ObjectScale S3 over RDMA engine.
- *
- * Reads the NIXL_OBJ_ENDPOINT_OVERRIDE environment variable to configure
- * the Dell ObjectScale endpoint. Tests are skipped if the variable is not set.
- * Requires cuobjclient library (HAVE_CUOBJ_CLIENT).
- */
-class setupObjDellTestFixture : public setupBackendTestFixture {
-protected:
-    setupObjDellTestFixture() {
-        const char *endpoint = std::getenv("NIXL_OBJ_ENDPOINT_OVERRIDE");
-        if (endpoint && endpoint[0] != '\0') {
-            obj_dell_params["endpoint_override"] = endpoint;
-            localBackendEngine_ = std::make_shared<nixlObjEngine>(&GetParam());
-        }
-    }
-
-    void
-    SetUp() override {
-        const char *endpoint = std::getenv("NIXL_OBJ_ENDPOINT_OVERRIDE");
-        if (!endpoint || endpoint[0] == '\0') {
-            GTEST_SKIP() << "NIXL_OBJ_ENDPOINT_OVERRIDE not set, skipping Dell tests";
-        }
-        setupBackendTestFixture::SetUp();
-    }
-};
-
-TEST_P(setupObjDellTestFixture, DellXferTest) {
-    transferHandler<DRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, dell_agent_name, dell_agent_name, false, 1);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-    transfer.resetLocalMem();
-    transfer.testTransfer(NIXL_READ);
-    transfer.checkLocalMem();
-}
-
-TEST_P(setupObjDellTestFixture, DellXferMultiBufsTest) {
-    transferHandler<DRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, dell_agent_name, dell_agent_name, false, 3);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-    transfer.resetLocalMem();
-    transfer.testTransfer(NIXL_READ);
-    transfer.checkLocalMem();
-}
-
-TEST_P(setupObjDellTestFixture, DellQueryMemTest) {
-    transferHandler<DRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, dell_agent_name, dell_agent_name, false, 3);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-
-    nixl_reg_dlist_t descs(OBJ_SEG);
-    descs.addDesc(nixlBlobDesc(nixlBasicDesc(), "test-obj-key-0"));
-    descs.addDesc(nixlBlobDesc(nixlBasicDesc(), "test-obj-key-1"));
-    descs.addDesc(nixlBlobDesc(nixlBasicDesc(), "test-obj-key-nonexistent"));
-    std::vector<nixl_query_resp_t> resp;
-    localBackendEngine_->queryMem(descs, resp);
-
-    EXPECT_EQ(resp.size(), 3);
-    EXPECT_EQ(resp[0].has_value(), true);
-    EXPECT_EQ(resp[1].has_value(), true);
-    EXPECT_EQ(resp[2].has_value(), false);
-}
-
-#ifdef HAVE_CUDA
-// GPU memory (VRAM_SEG) transfer test for Dell ObjectScale RDMA engine.
-// Exercises the VRAM-specific code paths: cuMemObjGetDescriptor/PutDescriptor for
-// RDMA descriptor registration, and putObjectRdmaAsync/getObjectRdmaAsync for
-// GPU-direct RDMA transfers.
-TEST_P(setupObjDellTestFixture, DellVramXferTest) {
-    transferHandler<VRAM_SEG, OBJ_SEG> transfer(
-        localBackendEngine_, localBackendEngine_, dell_agent_name, dell_agent_name, false, 1);
-    transfer.setLocalMem();
-    transfer.testTransfer(NIXL_WRITE);
-    transfer.resetLocalMem();
-    transfer.testTransfer(NIXL_READ);
-    transfer.checkLocalMem();
-}
-#endif // HAVE_CUDA
-
-INSTANTIATE_TEST_SUITE_P(ObjDellTests,
-                         setupObjDellTestFixture,
-                         testing::Values(obj_dell_test_params));
-
-#endif // HAVE_CUOBJ_CLIENT
 
 } // namespace gtest::plugins::obj
